Move payment methods out of Order class and into OrderComponents::Payment module
